+++ ProjetoPetMedDigital/ProjetoPetMedDigital/Models/Servico.cs
@@ -1,4 +1,4 @@
-using PetMed_Digital.Models;
+using PetMed_Digital.Models; // Adicionado para BaseModel
 using System;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
@@ -14,13 +14,13 @@ namespace ProjetoPetMedDigital.Models
         public string NomeServico { get; set; }
         public int IdVeterinario { get; set; }
         public DateTime Data { get; set; }
-        public DateTime Hora { get; set; }
-        public long Status { get; set; }
+        public DateTime Hora { get; set; } // Considerar TimeSpan ou apenas DateTime para Data completa
+        public int Status { get; set; } // Alterado de long para int
         public decimal PrecoVenda { get; set; }
         public string Descricao { get; set; }
         public int IdAgendamento { get; set; } // Chave estrangeira para Agendamento
-        public int IdProduto { get; set; } // Chave estrangeira para ItemEstoque (não estava no SQL, adicionei baseado no Procedimento e Vacina)
-        public int IdValor { get; set; } // Chave estrangeira para Valor
+        public int IdProduto { get; set; } // Chave estrangeira para ItemEstoque.IdProduto
+        public int IdValor { get; set; } // Chave estrangeira para Valor.IdValor
 
         // Propriedades de navegação
         public Agendamento Agendamento { get; set; }
